requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GitHub integration
PyGithub==2.1.1
requests==2.31.0

# Diff parsing
unidiff==0.7.5

# LangGraph for multi-agent orchestration
langgraph==0.0.62
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
httpx==0.25.2

# Development
python-dotenv==1.0.0

# Frontend (optional - only needed for Streamlit UI)
streamlit>=1.28.0